refactor: Organize queue factories and address test timestamp
This commit incorporates several refactoring changes and a test fix:

1.  **Timestamp Assertion Update in E2E Test:**
    *   In `tsercom/runtime_e2etest.py`, the comment regarding a potentially
        flaky timestamp assertion (`assert first.timestamp == stop_timestamp`)
        within the `__check_initialization` function was removed. The
        assertion itself remains, as `FakeRuntime` uses a global fixed
        timestamp.

2.  **`DefaultMultiprocessQueueFactory` Refactoring:**
    *   I moved `DefaultMultiprocessQueueFactory` from
        `tsercom/threading/multiprocess/multiprocess_queue_factory.py`
        to its own new file:
        `tsercom/threading/multiprocess/default_queue_factory.py`.
    *   I updated all relevant import statements in
        `tsercom/api/split_process/split_runtime_factory_factory.py` and
        the corresponding unit test file.
    *   I renamed the unit test file
        `tsercom/threading/multiprocess/multiprocess_queue_factory_unittest.py`
        to `tsercom/threading/multiprocess/default_queue_factory_unittest.py`
        to align with the new source file name.

3.  **`TorchMultiprocessQueueFactory` File Renaming:**
    *   I renamed `tsercom/threading/multiprocess/torch_queue_factory.py` to
        `tsercom/threading/multiprocess/torch_multiprocess_queue_factory.py`
        to better match the class name.
    *   I updated all relevant import statements in
        `tsercom/api/split_process/split_runtime_factory_factory.py` and
        the corresponding unit test file.
    *   I renamed the unit test file
        `tsercom/threading/multiprocess/torch_queue_factory_unittest.py`
        to `tsercom/threading/multiprocess/torch_multiprocess_queue_factory_unittest.py`.

These changes improve code organization and address a minor concern in
an E2E test. I still need to clean up the now-reduced
`multiprocess_queue_factory.py` (docstring update) and run full
linters/tests for this set of changes.
